pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "crewai>=0.121.1",
    "crewai-tools==0.46.0",
    "jupyterlab>=4.4.3",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-google-genai==2.1.5",
    "langchain-openai==0.2.14",
    "python-dotenv>=1.1.0",
    "ruff>=0.11.12",
]
